SQUASH:ipv6: Fix int addr_len pointer for older linux patch
Squashed Commit :
	* inet: fix addr_len/msg->msg_namelen assignment in recv_error and rxpmtu functions
	* net: ipv6: Amend ipv6_recv_error usage

(inet fix)
[ Upstream commit 85fbaa7 ]

Commit bceaa90 ("inet: prevent leakage
of uninitialized memory to user in recv syscalls") conditionally updated
addr_len if the msg_name is written to. The recv_error and rxpmtu
functions relied on the recvmsg functions to set up addr_len before.

As this does not happen any more we have to pass addr_len to those
functions as well and set it to the size of the corresponding sockaddr
length.

This broke traceroute and such.

(net fix)

Usage of ipv6_recv_error, included in commit f704820, needs to be
updated to account for the changes in ad25b5d.

Fixes:
  net/ipv6/ping.c: In function 'pingv6_init':
  net/ipv6/ping.c:84:29: warning: assignment from incompatible pointer type [enabled by default]

Fixes: bceaa90 ("inet: prevent leakage of uninitialized memory to user in recv syscalls")
Reported-by: Brad Spengler <spender@grsecurity.net>
Reported-by: Tom Labanowski
Cc: mpb <mpb.mail@gmail.com>
Cc: David S. Miller <davem@davemloft.net>
Cc: Eric Dumazet <eric.dumazet@gmail.com>
Signed-off-by: Hannes Frederic Sowa <hannes@stressinduktion.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Luis Cruz <ljc2491@gmail.com>
Signed-off-by: Dicky Herlambang <herlambangdicky5@gmail.com>